feat(login): 이메일 입력 및 유효성 검사 기능 구현
- 이메일 미입력 시 'ID를 입력해주세요.' 에러 메시지가 뜨도록 했습니다.
- 이메일 형식이 아닌 경우 'ID는 이메일 형식으로 입력해주세요.' 에러 메시지가 뜨도록 했습니다.
- 조건을 만족하면 에러 메시지 자동 제거되도록 실시간 onChange 검사를 적용했습니다.
- 에러 발생 시 메시지 뿐 아니라 입력창 하단 border 색상을 빨간색으로 강조하도록 했습니다.
- README에 구현 완료한 내용을 표시했습니다.
