+++ README.md
@@ -13,12 +13,12 @@
 - [ ] 본인만의 기준으로 일관된 코드를 작성해주세요.
 - [ ] 기능 단위로 나누어 커밋을 해주세요.
 - [ ] 로그인 기능을 만들어주세요.
-  - [ ] ID는 아래의 조건에 맞게 동작해요.
-    - [ ] ID는 반드시 입력 되어야 해요.
-    - [ ] ID는 이메일 형식이어야 해요.
-    - [ ] 만약 ID Input Field를 벗어났을 때 위 조건이 충족되지 않으면 에러 메시지를 보여줘요.  
+  - [x] ID는 아래의 조건에 맞게 동작해요.
+    - [x] ID는 반드시 입력 되어야 해요.
+    - [x] ID는 이메일 형식이어야 해요.
+    - [x] 만약 ID Input Field를 벗어났을 때 위 조건이 충족되지 않으면 에러 메시지를 보여줘요.  
            (ex. ID를 입력해주세요., ID는 이메일 형식으로 입력해주세요.)
-    - [ ] 위 조건이 충족되면 에러 메시지가 사라져요.
+    - [x] 위 조건이 충족되면 에러 메시지가 사라져요.
   - [ ] PW는 아래 조건에 맞게 동작해요.
     - [ ] PW는 최소 8글자 이상 입력되어야 해요.
     - [ ] 만약 PW Input Field를 벗어났을 때 위 조건이 충족되지 않으면 에러 메시지를 보여줘요.
